+++ src/main/java/net/minecraft/block/BlockDropper.java
@@ -38,7 +38,7 @@ public class BlockDropper extends BlockDispenser {
                 if (itemstack != null) {
                     EnumFacing enumfacing = (EnumFacing) worldIn.getBlockState(pos).getValue(FACING);
                     BlockPos blockpos = pos.offset(enumfacing);
-                    IInventory iinventory = TileEntityHopper.getInventoryAtPosition(worldIn, (double) blockpos.getX(), (double) blockpos.getY(), (double) blockpos.getZ());
+                    IInventory iinventory = TileEntityHopper.getInventoryAtPosition(worldIn, blockpos.getX(), blockpos.getY(), blockpos.getZ());
                     ItemStack itemstack1;
 
                     if (iinventory == null) {
